refactor: move create dukan components to a separate package
